metrics-server: try replacing helm with kustomize
This PR proposes the use of the kustomize included in `kubectl` instead of the third-party `helm` to deploy metrics-server.

The way this works is roughly:

1. create a DOKS cluster (just like before)
1. clone `kubernetes-incubator/metrics-server` at a specific git ref
1. call `kubectl kustomize` in the `metrics-server` directory
  * this makes use of the `kustomization.yaml` file in the `metrics-server` directory
  * the `kustomization.yaml` file points at example deployment manifests in the cloned `kubernetes-incubator/metrics-server` directory
  * the `kustomization.yaml` patches the upstream example deployment manifests with settings similar to what came with the `helm` chart
1. just like before, wait until the `metrics-server` pod(s) are ready

One major difference from the `helm` chart is that with `helm` it was easy to deploy `metrics-server` to its own namespace whereas the upstream example manifests on which the `kustomize` approach is based assume deployment in `kube-system`. It wasn't straightforward to use `kustomize` patches or bases to do something similar -- I'm sure it's possible, I just haven't found the minimal set of patches necessary yet.